requirements.txt
aniso8601==9.0.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-RESTful==0.3.9
future==0.18.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2==2.9.3
python-dotenv==0.19.2
python-whois==0.7.3
pytz==2021.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
urllib3==1.26.8
Werkzeug==2.0.2
